[Decomp][Spell] Cohesion split: extract SpellAuraPeriodic handlers to SpellAuraPeriodic.cpp
